Suggested fixes, removed old resource/service, fixed tests
