+++ server/planning/events/__init__.py
@@ -31,7 +31,6 @@ from .events_post import EventsPostService, EventsPostResource
 from .events_cancel import EventsCancelService, EventsCancelResource
 from .events_reschedule import EventsRescheduleService, EventsRescheduleResource
 from .events_postpone import EventsPostponeService, EventsPostponeResource
-from .events_update_time import EventsUpdateTimeService, EventsUpdateTimeResource
 from .events_update_repetitions import (
     EventsUpdateRepetitionsService,
     EventsUpdateRepetitionsResource,
@@ -106,15 +105,6 @@ def init_app(app):
     )
     EventsPostponeResource(EventsPostponeResource.endpoint_name, app=app, service=events_postpone_service)
 
-    events_update_time_service = EventsUpdateTimeService(
-        EventsUpdateTimeResource.endpoint_name, backend=superdesk.get_backend()
-    )
-    EventsUpdateTimeResource(
-        EventsUpdateTimeResource.endpoint_name,
-        app=app,
-        service=events_update_time_service,
-    )
-
     events_update_repetitions_service = EventsUpdateRepetitionsService(
         EventsUpdateRepetitionsResource.endpoint_name, backend=superdesk.get_backend()
     )
@@ -143,6 +133,7 @@ def init_app(app):
 
     # listen to async signals
     signals.events_created.connect(events_history_service.on_item_created)
+    signals.event_time_updated.connect(events_history_service.on_item_created)
 
     app.on_updated_events += events_history_service.on_item_updated
 
@@ -156,7 +147,6 @@ def init_app(app):
     app.on_updated_events_cancel += events_history_service.on_cancel
     app.on_updated_events_reschedule += events_history_service.on_reschedule
     app.on_updated_events_postpone += events_history_service.on_postpone
-    app.on_updated_events_update_time += events_history_service.on_update_time
     app.on_locked_events += events_search_service.on_locked_event
 
     # Privileges
